Fix: install `requests` properly during setup
Setup was failing on a fresh virtualenv because setup.py
imported tentapp, which imported requests, which wasn't
installed yet (even though requests is an install_requires,
it gets installed *after* tentapp).

I worked around this by hardcoding the version into
setup.py. The version is now stored in both setup.py and
tentapp/__init__.py and must be updated in both places.
Sorry that it violates DRY, but it does fix the setup
process and looking at Flask (for example) it too
duplicates the version number in this way.

FYI, one thing I tried that did NOT work was adding a
setup_requires for requests as well as the
install_requires. I'm not entirely sure why that doesn't
work, but some webpage somewhere claims pip doesn't support
setup_requires when installing from bundles, so there you
go.